Split code up in 2 diferent functions, one to populate and other one to add likes
